Added CellHighlighter class
